requirements.txt
flask
flask-cors
gunicorn
google-cloud-translate
gruut
gruut-lang-en